+++ nailmirror/src/pages/try-on-ar/index.js
@@ -83,7 +83,7 @@ Page({
         hdUrl: hd.hdUrl
       });
       wx.hideLoading();
-      wx.navigateTo({ url: '/pages/hd-output/index?styleId=' + this.data.styleId + '&hdUrl=' + encodeURIComponent(hd.hdUrl) });
+      require('../../utils/hd-output-nav').navigateTo(this.data.styleId, hd.hdUrl);
     } catch (e) {
       wx.hideLoading();
       wx.showToast({ title: '出片失败', icon: 'none' });